cleanup - removing a duplciate
