+++ sdk/runanywhere-swift/Sources/RunAnywhere/Public/Models/Voice/ModularPipelineEvent.swift
@@ -1,5 +1,13 @@
 import Foundation
 
+/// Stages in the voice pipeline
+public enum PipelineStage: String, CaseIterable {
+    case vad = "VAD"
+    case transcription = "Speech-to-Text"
+    case llmGeneration = "LLM Generation"
+    case textToSpeech = "Text-to-Speech"
+}
+
 /// Extended events for modular voice pipeline processing
 public enum ModularPipelineEvent {
     // VAD events
@@ -40,4 +48,36 @@ public enum ModularPipelineEvent {
     case pipelineStarted
     case pipelineError(Error)
     case pipelineCompleted
+}
+
+/// Complete result from voice pipeline
+public struct VoicePipelineResult {
+    /// The transcription result from STT
+    public let transcription: VoiceTranscriptionResult
+
+    /// The LLM generated response text
+    public let llmResponse: String
+
+    /// The synthesized audio output (if TTS enabled)
+    public let audioOutput: Data?
+
+    /// Total processing time
+    public let processingTime: TimeInterval
+
+    /// Per-stage timing metrics
+    public let stageTiming: [PipelineStage: TimeInterval]
+
+    public init(
+        transcription: VoiceTranscriptionResult,
+        llmResponse: String,
+        audioOutput: Data? = nil,
+        processingTime: TimeInterval = 0,
+        stageTiming: [PipelineStage: TimeInterval] = [:]
+    ) {
+        self.transcription = transcription
+        self.llmResponse = llmResponse
+        self.audioOutput = audioOutput
+        self.processingTime = processingTime
+        self.stageTiming = stageTiming
+    }
 }
